+++ tests/mock_vws/fixtures/vuforia_backends.py
@@ -22,7 +22,10 @@ from mock_vws._flask_server.vws import VWS_FLASK_APP
 from mock_vws.database import CloudDatabase, VuMarkDatabase
 from mock_vws.states import States
 from mock_vws.target import VuMarkTarget
-from tests.mock_vws.fixtures.credentials import VuMarkCloudDatabase
+from tests.mock_vws.fixtures.credentials import (
+    InactiveVuMarkCloudDatabase,
+    VuMarkCloudDatabase,
+)
 from tests.mock_vws.utils.retries import RETRY_ON_TOO_MANY_REQUESTS
 
 LOGGER = logging.getLogger(name=__name__)
@@ -72,11 +75,16 @@ def _vumark_database(
         name="mock-vumark-target",
         target_id=vumark_vuforia_database.target_id,
     )
+    processing_target = VuMarkTarget(
+        name="mock-processing-vumark-target",
+        target_id=vumark_vuforia_database.processing_target_id,
+        processing_time_seconds=9999,
+    )
     return VuMarkDatabase(
         database_name=vumark_vuforia_database.target_manager_database_name,
         server_access_key=vumark_vuforia_database.server_access_key,
         server_secret_key=vumark_vuforia_database.server_secret_key,
-        vumark_targets={vumark_target},
+        vumark_targets={vumark_target, processing_target},
     )
 
 
@@ -84,14 +92,16 @@ def _vumark_database(
 def _enable_use_real_vuforia(
     *,
     working_database: CloudDatabase,
-    inactive_database: CloudDatabase,
+    inactive_cloud_database: CloudDatabase,
     vumark_vuforia_database: VuMarkCloudDatabase,
+    inactive_vumark_database: InactiveVuMarkCloudDatabase,
     monkeypatch: pytest.MonkeyPatch,
 ) -> Generator[None]:
     """Test against the real Vuforia."""
     assert monkeypatch
-    assert inactive_database
+    assert inactive_cloud_database
     assert vumark_vuforia_database
+    assert inactive_vumark_database
     _delete_all_targets(database_keys=working_database)
     yield
 
@@ -100,8 +110,9 @@ def _enable_use_real_vuforia(
 def _enable_use_mock_vuforia(
     *,
     working_database: CloudDatabase,
-    inactive_database: CloudDatabase,
+    inactive_cloud_database: CloudDatabase,
     vumark_vuforia_database: VuMarkCloudDatabase,
+    inactive_vumark_database: InactiveVuMarkCloudDatabase,
     monkeypatch: pytest.MonkeyPatch,
 ) -> Generator[None]:
     """Test against the in-memory mock Vuforia."""
@@ -114,22 +125,29 @@ def _enable_use_mock_vuforia(
         client_secret_key=working_database.client_secret_key,
     )
 
-    inactive_database = CloudDatabase(
+    inactive_cloud_database = CloudDatabase(
         state=States.PROJECT_INACTIVE,
-        database_name=inactive_database.database_name,
-        server_access_key=inactive_database.server_access_key,
-        server_secret_key=inactive_database.server_secret_key,
-        client_access_key=inactive_database.client_access_key,
-        client_secret_key=inactive_database.client_secret_key,
+        database_name=inactive_cloud_database.database_name,
+        server_access_key=inactive_cloud_database.server_access_key,
+        server_secret_key=inactive_cloud_database.server_secret_key,
+        client_access_key=inactive_cloud_database.client_access_key,
+        client_secret_key=inactive_cloud_database.client_secret_key,
     )
     vumark_database = _vumark_database(
         vumark_vuforia_database=vumark_vuforia_database,
     )
+    inactive_vumark_db = VuMarkDatabase(
+        state=States.PROJECT_INACTIVE,
+        database_name=inactive_vumark_database.target_manager_database_name,
+        server_access_key=inactive_vumark_database.server_access_key,
+        server_secret_key=inactive_vumark_database.server_secret_key,
+    )
 
     with MockVWS() as mock:
         mock.add_cloud_database(cloud_database=working_database)
-        mock.add_cloud_database(cloud_database=inactive_database)
+        mock.add_cloud_database(cloud_database=inactive_cloud_database)
         mock.add_vumark_database(vumark_database=vumark_database)
+        mock.add_vumark_database(vumark_database=inactive_vumark_db)
         yield
 
 
@@ -137,8 +155,9 @@ def _enable_use_mock_vuforia(
 def _enable_use_docker_in_memory(
     *,
     working_database: CloudDatabase,
-    inactive_database: CloudDatabase,
+    inactive_cloud_database: CloudDatabase,
     vumark_vuforia_database: VuMarkCloudDatabase,
+    inactive_vumark_database: InactiveVuMarkCloudDatabase,
     monkeypatch: pytest.MonkeyPatch,
 ) -> Generator[None]:
     """Test against mock Vuforia created to be run in a container."""
@@ -165,7 +184,12 @@ def _enable_use_docker_in_memory(
     vumark_database = _vumark_database(
         vumark_vuforia_database=vumark_vuforia_database,
     )
-    (vumark_target,) = vumark_database.vumark_targets
+    inactive_vumark_db = VuMarkDatabase(
+        state=States.PROJECT_INACTIVE,
+        database_name=inactive_vumark_database.target_manager_database_name,
+        server_access_key=inactive_vumark_database.server_access_key,
+        server_secret_key=inactive_vumark_database.server_secret_key,
+    )
 
     with responses.RequestsMock(assert_all_requests_are_fired=False) as mock:
         add_flask_app_to_mock(
@@ -211,7 +235,7 @@ def _enable_use_docker_in_memory(
         )
         requests.post(
             url=cloud_databases_url,
-            json=inactive_database.to_dict(),
+            json=inactive_cloud_database.to_dict(),
             timeout=30,
         )
         requests.post(
@@ -220,13 +244,19 @@ def _enable_use_docker_in_memory(
             timeout=30,
         )
         requests.post(
-            url=(
-                f"{vumark_databases_url}"
-                f"/{vumark_database.database_name}/vumark_targets"
-            ),
-            json=vumark_target.to_dict(),
+            url=vumark_databases_url,
+            json=inactive_vumark_db.to_dict(),
             timeout=30,
         )
+        for vumark_target in vumark_database.vumark_targets:
+            requests.post(
+                url=(
+                    f"{vumark_databases_url}"
+                    f"/{vumark_database.database_name}/vumark_targets"
+                ),
+                json=vumark_target.to_dict(),
+                timeout=30,
+            )
 
         yield
 
@@ -290,8 +320,9 @@ def fixture_verify_mock_vuforia(
     *,
     request: pytest.FixtureRequest,
     vuforia_database: CloudDatabase,
-    inactive_database: CloudDatabase,
+    inactive_cloud_database: CloudDatabase,
     vumark_vuforia_database: VuMarkCloudDatabase,
+    inactive_vumark_database: InactiveVuMarkCloudDatabase,
     monkeypatch: pytest.MonkeyPatch,
 ) -> Generator[None]:
     """Test functions which use this fixture are run multiple times. Once
@@ -318,8 +349,9 @@ def fixture_verify_mock_vuforia(
 
     yield from enable_function(
         working_database=vuforia_database,
-        inactive_database=inactive_database,
+        inactive_cloud_database=inactive_cloud_database,
         vumark_vuforia_database=vumark_vuforia_database,
+        inactive_vumark_database=inactive_vumark_database,
         monkeypatch=monkeypatch,
     )
 
@@ -337,8 +369,9 @@ def mock_only_vuforia(
     *,
     request: pytest.FixtureRequest,
     vuforia_database: CloudDatabase,
-    inactive_database: CloudDatabase,
+    inactive_cloud_database: CloudDatabase,
     vumark_vuforia_database: VuMarkCloudDatabase,
+    inactive_vumark_database: InactiveVuMarkCloudDatabase,
     monkeypatch: pytest.MonkeyPatch,
 ) -> Generator[None]:
     """Test functions which use this fixture are run multiple times. Once
@@ -365,7 +398,8 @@ def mock_only_vuforia(
 
     yield from enable_function(
         working_database=vuforia_database,
-        inactive_database=inactive_database,
+        inactive_cloud_database=inactive_cloud_database,
         vumark_vuforia_database=vumark_vuforia_database,
+        inactive_vumark_database=inactive_vumark_database,
         monkeypatch=monkeypatch,
     )